==> chapter7航空公司客户价值分析/code/04-zscore_data.py <==
# 代码7-7

# 属性选择、构造与数据标准化

import pandas as pd
import numpy as np

# 读取数据清洗后的数据
cleanedfile = '../tmp/data_cleaned.csv'  # 数据清洗后保存的文件路径
airline = pd.read_csv(cleanedfile, encoding='utf-8')
# 选取需求属性
airline_selection = airline[['FFP_DATE', 'LOAD_TIME', 'LAST_TO_END',
                                     'FLIGHT_COUNT', 'SEG_KM_SUM', 'avg_discount']]
print('筛选的属性前5行为：\n', airline_selection.head())



# 代码7-8

# 构造属性L
L = pd.to_datetime(airline_selection['LOAD_TIME']) - \
    pd.to_datetime(airline_selection['FFP_DATE'])
L = L.astype('str').str.split().str[0]
L = L.astype('int')/30

# 合并属性
airline_features = pd.concat([L, airline_selection.iloc[:, 2:]], axis=1)
airline_features.columns = ['L', 'R', 'F', 'M', 'C']
print('构建的LRFMC属性前5行为：\n', airline_features.head())

# 数据标准化
from sklearn.preprocessing import StandardScaler
data = StandardScaler().fit_transform(airline_features)
np.savez('../tmp/airline_scale.npz', data)
# np.savetxt('../tmp/airline_scale.csv', data)
print('标准化后LRFMC五个属性为：\n', data[:5, :])
