==> chapter7航空公司客户价值分析/code/05-KMeans_cluster.py <==
# 代码7-9

# K-means聚类

import pandas as pd
import numpy as np
from sklearn.cluster import KMeans  # 导入kmeans算法

# 读取标准化后的数据
airline_scale = np.load('../tmp/airline_scale.npz')['arr_0']
k = 5  # 确定聚类中心数

# 构建模型，随机种子设为123
kmeans_model = KMeans(n_clusters=k, n_jobs=4, random_state=123)
fit_kmeans = kmeans_model.fit(airline_scale)  # 模型训练

# 查看聚类结果
kmeans_cc = kmeans_model.cluster_centers_  # 聚类中心
print('各类聚类中心为：\n', kmeans_cc)
kmeans_labels = kmeans_model.labels_  # 样本的类别标签
print('各样本的类别标签为：\n', kmeans_labels)
r1 = pd.Series(kmeans_model.labels_).value_counts()  # 统计不同类别样本的数目
print('最终每个类别的数目为：\n', r1)
# 输出聚类分群的结果
cluster_center = pd.DataFrame(kmeans_model.cluster_centers_,
                              columns=['ZL', 'ZR', 'ZF', 'ZM', 'ZC'])   # 将聚类中心放在数据框中
# cluster_center.index = pd.DataFrame(kmeans_model.labels_).\
#                   drop_duplicates().iloc[:, 0]  # 将样本类别作为数据框索引



# 代码7-10

# %matplotlib inline
import matplotlib.pyplot as plt

# 客户分群雷达图
labels = ['ZL', 'ZR', 'ZF', 'ZM', 'ZC']
legen = ['客户群' + str(i + 1) for i in cluster_center.index]  # 客户群命名，作为雷达图的图例
lstype = ['-', '--', (0, (3, 5, 1, 5, 1, 5)), ':', '-.']

# 由于雷达图要保证数据闭合，因此再添加ZL列，并转换为 np.ndarray
centers = pd.concat([cluster_center, cluster_center[['ZL']]], axis=1).values

# 分割圆周长，并让其闭合
n = len(legen)
angle = np.linspace(0, 2 * np.pi, n, endpoint=False)

angles = np.concatenate([angle, [angle[0]]])

# 绘图
fig = plt.figure(figsize=(8, 6))
ax = fig.add_subplot(111, polar=True)  # 以极坐标的形式绘制图形
plt.rcParams['font.sans-serif'] = ['SimHei']  # 用来正常显示中文标签
plt.rcParams['axes.unicode_minus'] = False  # 用来正常显示负号 
# 画线
for i in range(n):
    ax.plot(angles, centers[i], linestyle=lstype[i], linewidth=2)
# 添加属性标签
ax.set_thetagrids(angle * 180 / np.pi, labels)
plt.title('客户特征分析雷达图')
plt.legend(legen)
plt.show()
plt.close
